Revert "add as much encoding logic as possible into scm-slang"
This reverts commit cda6f1a2ee228bd5efb4d26c5b3ae4c49eaf5f08.
